+++ src/main/groovy/life/qbic/datamodel/dtos/business/Offer.groovy
@@ -45,17 +45,17 @@ class Offer {
      */
     final double totalPrice
     /**
-     * The net value of the offer price
+     * The net price of all items without taxes and overhead
      */
-    final double net
+    final double netPrice
     /**
-     * The overheads applied to the offer price
+     * The amount of taxes, part of the total total price
      */
-    final double overhead
+    final double taxes
     /**
-     * The value added taxes for the offer price
+     * The amount of overheads part, of the total price
      */
-    final double vat
+    final double overheads
     /**
      * The identifier for the offer which makes it distinguishable from other offers
      */
@@ -67,19 +67,30 @@ class Offer {
 
     static class Builder {
 
-        Date modificationDate
-        Date expirationDate
-        Customer customer
-        ProjectManager projectManager
+        /*
+        Overall offer describing properties
+         */
         String projectTitle
         String projectDescription
-        List<ProductItem> items
-        double totalPrice
-        double vat
-        double net
-        double overhead
-        OfferId identifier
+        Customer customer
+        ProjectManager projectManager
         Affiliation selectedCustomerAffiliation
+        Date modificationDate
+        Date expirationDate
+        List<ProductItem> items
+
+        /*
+        Offer identifier
+         */
+        OfferId identifier
+
+        /*
+        Price related properties
+         */
+        double totalPrice
+        double netPrice
+        double taxes
+        double overheads
 
         Builder(Customer customer, ProjectManager projectManager, String projectTitle, String projectDescription, Affiliation selectedCustomerAffiliation) {
             this.customer = Objects.requireNonNull(customer, "Customer must not be null")
@@ -88,6 +99,10 @@ class Offer {
             this.projectDescription = Objects.requireNonNull(projectDescription, "Project Description must not be null")
             this.selectedCustomerAffiliation = Objects.requireNonNull(selectedCustomerAffiliation, "Customer Affiliation must not be null")
             this.items = []
+            netPrice = 0
+            overheads = 0
+            taxes = 0
+            totalPrice = 0
         }
 
         Builder modificationDate(Date modificationDate) {
@@ -97,6 +112,21 @@ class Offer {
 
         Builder expirationDate(Date expirationDate) {
             this.expirationDate = expirationDate
+            return this
+        }
+
+        Builder netPrice(Double netPrice) {
+            this.netPrice = netPrice
+            return this
+        }
+
+        Builder overheads(Double overheads) {
+            this.overheads = overheads
+            return this
+        }
+
+        Builder taxes(Double taxes) {
+            this.taxes = taxes
             return this
         }
 
@@ -143,12 +173,12 @@ class Offer {
         this.projectDescription = builder.projectDescription
         this.projectTitle = builder.projectTitle
         this.items = builder.items
-        this.totalPrice = builder.totalPrice
-        this.vat = builder.vat
-        this.net = builder.net
-        this.overhead = builder.overhead
         this.identifier = builder.identifier
         this.selectedCustomerAffiliation = builder.selectedCustomerAffiliation
+        this.netPrice = builder.netPrice
+        this.overheads = builder.overheads
+        this.taxes = builder.taxes
+        this.totalPrice = builder.totalPrice
     }
 
     /**