+++ src/main/groovy/life/qbic/datamodel/dtos/business/services/AtomicProduct.groovy
@@ -1,5 +1,7 @@
 package life.qbic.datamodel.dtos.business.services
 
+import groovy.transform.EqualsAndHashCode
+
 /**
  * Describes a product type that can only have positive natural unit multipliers (N={0,1...,inf+})
  *
@@ -8,6 +10,7 @@ package life.qbic.datamodel.dtos.business.services
  *
  * @since 1.12.0
  */
+@EqualsAndHashCode(callSuper = true)
 class AtomicProduct extends Product {
   /**
    * Basic product constructor.