+++ CHANGELOG.rst
@@ -10,7 +10,7 @@ This project adheres to `Semantic Versioning <https://semver.org/>`_.
 
 **Added**
 * Introduce profiles to pom to differentiate between OSGI and non-OSGI packaging
-
+* Added a netPrice, taxes and overheads property to the offer DTO
 * Add bnd.bnd file to enable OSGI bundling
 
 **Fixed**